Admin can only edit user details, not other admins

+++ code/SmartLocker - Backend/SmartLocker/src/main/java/com/group17/SmartLocker/service/user/UserService.java
@@ -4,7 +4,9 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.group17.SmartLocker.dto.LockerLogDto;
 import com.group17.SmartLocker.dto.UserDetailsDto;
+import com.group17.SmartLocker.enums.Role;
 import com.group17.SmartLocker.exception.ResourceNotFoundException;
+import com.group17.SmartLocker.exception.UnauthorizedActionException;
 import com.group17.SmartLocker.model.LockerLog;
 import com.group17.SmartLocker.model.User;
 import com.group17.SmartLocker.repository.UserRepository;
@@ -62,6 +64,7 @@ public class UserService implements IUserService {
 
     @Override
     public UserDetailsDto getUserById(String id){
+
         User user = userRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("User not exist with id: " + id));
 
@@ -82,16 +85,26 @@ public class UserService implements IUserService {
 
     @Override
     public User updateUser(String id, User userDetails){
+
         User updateUser = userRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("User not exist with id: " + id));
 
-        updateUser.setUsername(userDetails.getUsername());
-        updateUser.setFirstName(userDetails.getFirstName());
-        updateUser.setLastName(userDetails.getLastName());
-        updateUser.setContactNumber(userDetails.getContactNumber());
-        updateUser.setEmail(userDetails.getEmail());
 
-        return userRepository.save(updateUser);
+        if(updateUser.getRole().equals(Role.USER)){
+
+            // Admins can only update the regular users details , not other Admins
+            updateUser.setUsername(userDetails.getUsername());
+            updateUser.setFirstName(userDetails.getFirstName());
+            updateUser.setLastName(userDetails.getLastName());
+            updateUser.setContactNumber(userDetails.getContactNumber());
+            updateUser.setEmail(userDetails.getEmail());
+
+            return userRepository.save(updateUser);
+        }
+        else{
+            throw new UnauthorizedActionException("You are not permitted to this Action");
+        }
+
     }
 
     @Override
@@ -99,17 +112,26 @@ public class UserService implements IUserService {
         User user = userRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("User not found with id: " + id));
 
-        Set<String> allowedFields = Set.of("firstName", "lastName", "email", "contactNumber");
+        if(user.getRole().equals(Role.USER)){
 
-        updates.forEach((key, value) -> {
-            if (allowedFields.contains(key)){
-                Field field = ReflectionUtils.findField(User.class, key);
-                if (field != null) {
-                    field.setAccessible(true);
-                    ReflectionUtils.setField(field, user, value);
+            // Admins can only edit the details of the regular users, not other Admins
+
+            Set<String> allowedFields = Set.of("firstName", "lastName", "email", "contactNumber");
+
+            updates.forEach((key, value) -> {
+                if (allowedFields.contains(key)){
+                    Field field = ReflectionUtils.findField(User.class, key);
+                    if (field != null) {
+                        field.setAccessible(true);
+                        ReflectionUtils.setField(field, user, value);
+                    }
                 }
-            }
-        });
+            });
+
+        }
+        else{
+            throw new UnauthorizedActionException("You are not permitted to this Action");
+        }
 
         return userRepository.save(user);
     }
@@ -118,7 +140,15 @@ public class UserService implements IUserService {
     public void deleteUser(String id) {
         User user = userRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("User not exist with id: " + id));
-        userRepository.delete(user);
+
+        if(user.getRole().equals(Role.USER)){
+
+            // Admins can only delete the regular users, not other Admins
+            userRepository.delete(user);
+        }
+        else{
+            throw new UnauthorizedActionException("You are not permitted to this Action");
+        }
     }
 
     @Override
